unitmaker.go
package main

import (
    "flag"
    "fmt"
    "go/ast"
    "go/parser"
    "go/token"
    "os"
    "strings"
)

func main() {
    // Parsing command line argument
    flag.Parse()
    args := flag.Args()

    if len(args) != 1 {
        fmt.Println("Please provide a Go file as an argument")
        os.Exit(1)
    }

    filename := args[0]
    fset := token.NewFileSet()

    // Parsing the Go file
    node, err := parser.ParseFile(fset, filename, nil, parser.ParseComments)
    if err != nil {
        fmt.Println("Error parsing file:", err)
        os.Exit(1)
    }

    // Create the test file
    testFilename := strings.TrimSuffix(filename, ".go") + "_test.go"
    testFile, err := os.Create(testFilename)
    if err != nil {
        fmt.Println("Error creating test file:", err)
        os.Exit(1)
    }
    defer testFile.Close()

    // Write package declaration to test file
    fmt.Fprintf(testFile, "package %s\n\nimport \"testing\"\n\n", node.Name.Name)

    // Iterate through the AST and find functions and type declarations
    for _, f := range node.Decls {
        switch decl := f.(type) {
        case *ast.FuncDecl:
            generateTestForFunc(decl, testFile)
        case *ast.GenDecl:
            if decl.Tok == token.TYPE {
                for _, spec := range decl.Specs {
                    if typeSpec, ok := spec.(*ast.TypeSpec); ok {
                        generateTestForType(typeSpec, testFile)
                    }
                }
            }
        }
    }
}

func generateTestForFunc(fn *ast.FuncDecl, testFile *os.File) {
    // Generate a basic test function
    testName := "Test" + fn.Name.Name
    fmt.Fprintf(testFile, "// %s tests the %s function.\n", testName, fn.Name.Name)
    fmt.Fprintf(testFile, "func %s(t *testing.T) {\n", testName)

    // Mock inputs for parameters (using zero values for simplicity)
    if fn.Type.Params != nil {
        for _, p := range fn.Type.Params.List {
            for _, n := range p.Names {
                fmt.Fprintf(testFile, "    var %s %s // TODO: Add appropriate mock value\n", n.Name, exprString(p.Type))
            }
        }
    }

    // Call the function
    fmt.Fprintf(testFile, "    ")
    if fn.Type.Results != nil && len(fn.Type.Results.List) > 0 {
        fmt.Fprintf(testFile, "got := ")
    }
    fmt.Fprintf(testFile, "%s(", fn.Name.Name)
    if fn.Type.Params != nil {
        for i, p := range fn.Type.Params.List {
            for j, n := range p.Names {
                fmt.Fprintf(testFile, n.Name)
                if j < len(p.Names)-1 || i < len(fn.Type.Params.List)-1 {
                    fmt.Fprintf(testFile, ", ")
                }
            }
        }
    }
    fmt.Fprintf(testFile, ")\n")

    // Add a basic assertion
    fmt.Fprintf(testFile, "    // TODO: Add assertions based on the expected behavior of the function\n")
    
    fmt.Fprintf(testFile, "}\n\n")
}

func generateTestForType(typeSpec *ast.TypeSpec, testFile *os.File) {
    // Placeholder for generating tests for methods of a type
    fmt.Fprintf(testFile, "// TODO: Generate tests for methods of type %s\n", typeSpec.Name.Name)
}

func exprString(e ast.Expr) string {
    switch v := e.(type) {
    case *ast.Ident:
        return v.Name
    case *ast.ArrayType:
        return "[]" + exprString(v.Elt)
    case *ast.StarExpr:
        return "*" + exprString(v.X)
    case *ast.SelectorExpr:
        return exprString(v.X) + "." + v.Sel.Name
    // Add more cases as necessary for different types
    default:
        return fmt.Sprintf("%v", v)
    }
}
