unitmaker_test.go
package main

import "testing"

// Testmain tests the main function.
func Testmain(t *testing.T) {
    main()
    // TODO: Add assertions based on the expected behavior of the function
}

// TestgenerateTestForFunc tests the generateTestForFunc function.
func TestgenerateTestForFunc(t *testing.T) {
    var fn *ast.FuncDecl // TODO: Add appropriate mock value
    var testFile *os.File // TODO: Add appropriate mock value
    generateTestForFunc(fn, testFile)
    // TODO: Add assertions based on the expected behavior of the function
}

// TestgenerateTestForType tests the generateTestForType function.
func TestgenerateTestForType(t *testing.T) {
    var typeSpec *ast.TypeSpec // TODO: Add appropriate mock value
    var testFile *os.File // TODO: Add appropriate mock value
    generateTestForType(typeSpec, testFile)
    // TODO: Add assertions based on the expected behavior of the function
}

// TestexprString tests the exprString function.
func TestexprString(t *testing.T) {
    var e ast.Expr // TODO: Add appropriate mock value
    got := exprString(e)
    // TODO: Add assertions based on the expected behavior of the function
}

